requirements.txt
fastapi
uvicorn
pymongo
snowflake-connector-python
pandas
dash
plotly
ydata-profiling